source/requirement.txt
Django==1.11.23
django-cors-headers==3.2.0
djangorestframework==3.10.3
psycopg2-binary==2.7.5
pytz==2019.3
